Revert "CB Skills update"
