+++ Sources/oazix/CustomBehaviors/skillbars/ritualist_soul_twisting.py
@@ -20,7 +20,6 @@ from Sources.oazix.CustomBehaviors.skills.generic.keep_self_effect_up_utility im
 from Sources.oazix.CustomBehaviors.skills.generic.protective_spirit_utility import ProtectiveSpiritUtility
 from Sources.oazix.CustomBehaviors.skills.monk.strength_of_honor_utility import StrengthOfHonorUtility
 from Sources.oazix.CustomBehaviors.skills.paragon.fall_back_utility import FallBackUtility
-from Sources.oazix.CustomBehaviors.skills.ranger.energizing_wind_utility import EnergizingWindUtility
 from Sources.oazix.CustomBehaviors.skills.ritualist.armor_of_unfeeling_utility import ArmorOfUnfeelingUtility
 from Sources.oazix.CustomBehaviors.skills.ritualist.summon_spirit_utility import SummonSpiritUtility
 
@@ -41,7 +40,6 @@ class RitualistSoulTwisting_UtilitySkillBar(CustomBehaviorBaseUtility):
         self.armor_of_unfeeling_utility: CustomSkillUtilityBase = ArmorOfUnfeelingUtility(event_bus=self.event_bus, current_build=in_game_build, score_definition=ScoreStaticDefinition(80))
 
         # optional
-        self.energizing_wind_utility: CustomSkillUtilityBase = EnergizingWindUtility(event_bus=self.event_bus, current_build=in_game_build, score_definition=ScoreStaticDefinition(97))
         self.breath_of_the_great_dwarf_utility: CustomSkillUtilityBase = BreathOfTheGreatDwarfUtility(event_bus=self.event_bus, current_build=in_game_build, score_definition=ScorePerHealthGravityDefinition(0))
 
         # common
@@ -62,7 +60,6 @@ class RitualistSoulTwisting_UtilitySkillBar(CustomBehaviorBaseUtility):
             self.summon_spirit_kurzick,
             self.summon_spirit_luxon,
             self.armor_of_unfeeling_utility,
-            self.energizing_wind_utility,
             self.breath_of_the_great_dwarf_utility,
             self.ebon_vanguard_assassin_support,
             self.ebon_battle_standard_of_wisdom,
